renaming test files and fixing groups
