--- src/test/java/com/wikia/webdriver/testcases/discussions/PostDetails.java
@@ -1,70 +1,0 @@
-package com.wikia.webdriver.testcases.discussions;
-
-import com.wikia.webdriver.common.core.Assertion;
-import com.wikia.webdriver.common.core.annotations.Execute;
-import com.wikia.webdriver.common.core.annotations.InBrowser;
-import com.wikia.webdriver.common.core.helpers.User;
-import com.wikia.webdriver.common.templates.NewTestTemplate;
-import com.wikia.webdriver.pageobjectsfactory.pageobject.discussions.PostDetailsPage;
-
-import org.testng.annotations.Test;
-
-@Test(groups = {"Discussions", "PostDetails"})
-public class PostDetails extends NewTestTemplate {
-
-  private static final String DESKTOP_RESOLUTION = "1366x768";
-  private static final String MOBILE_RESOLUTION = "600x800";
-
-  /**
-   * ANONS ON MOBILE SECTION
-   */
-
-  @Test
-  @Execute(asUser = User.ANONYMOUS)
-  @InBrowser(browserSize = MOBILE_RESOLUTION)
-  public void anonUserOnMobileCanSeePostDetailsList() {
-    postDetailsListLoads();
-  }
-
-  /**
-   * ANONS ON DESKTOP SECTION
-   */
-
-  @Test
-  @Execute(asUser = User.ANONYMOUS)
-  @InBrowser(browserSize = DESKTOP_RESOLUTION)
-  public void anonUserOnDesktopCanSeePostDetailsList() {
-    postDetailsListLoads();
-  }
-
-  /**
-   * LOGGED IN USERS ON MOBILE SECTION
-   */
-
-  @Test
-  @Execute(asUser = User.USER_3)
-  @InBrowser(browserSize = MOBILE_RESOLUTION)
-  public void loggedInUserOnMobileCanSeePostDetailsList() {
-    postDetailsListLoads();
-  }
-
-  /**
-   * LOGGED IN USERS ON DESKTOP SECTION
-   */
-
-  @Test
-  @Execute(asUser = User.USER_3)
-  @InBrowser(browserSize = DESKTOP_RESOLUTION)
-  public void loggedInUserOnDesktopCanSeePostDetailsList() {
-    postDetailsListLoads();
-  }
-
-  /**
-   * TESTING METHODS SECTION
-   */
-
-  public void postDetailsListLoads() {
-    PostDetailsPage postDetails = new PostDetailsPage(driver).open();
-    Assertion.assertFalse(postDetails.isPostDetailsListEmpty());
-  }
-}
