+++ src/test/java/com/wikia/webdriver/testcases/discussions/Sharing.java
@@ -3,10 +3,9 @@ package com.wikia.webdriver.testcases.discussions;
 import com.wikia.webdriver.common.core.Assertion;
 import com.wikia.webdriver.common.core.annotations.Execute;
 import com.wikia.webdriver.common.core.annotations.InBrowser;
-import com.wikia.webdriver.common.core.configuration.Configuration;
 import com.wikia.webdriver.common.core.helpers.User;
-import com.wikia.webdriver.common.properties.Credentials;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
+import com.wikia.webdriver.pageobjectsfactory.pageobject.discussions.PostDetailsPage;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.discussions.PostsListPage;
 
 import org.testng.annotations.Test;
@@ -14,20 +13,17 @@ import org.testng.annotations.Test;
 /**
  * @ownership Content X-Wing Wikia
  */
-
-@Test(groups = {"Discussions", "PostListSharing"})
-public class PostsListSharing extends NewTestTemplate {
+@Test(groups = {"Discussions", "Sharing"})
+public class Sharing extends NewTestTemplate {
 
   private static final String DESKTOP_RESOLUTION = "1366x768";
   private static final String MOBILE_RESOLUTION = "600x800";
 
-  private Credentials credentials = Configuration.getCredentials();
-
   /**
-   * ANONYMOUS USER SECTION
+   * ANONYMOUS USER
    */
 
-  @Test
+  @Test(groups = {"Sharing_001"})
   @Execute(asUser = User.ANONYMOUS)
   @InBrowser(browserSize = MOBILE_RESOLUTION)
   public void socialNetworkIconsAreDisplayedToAnonymousUserOnMobile() {
@@ -35,25 +31,39 @@ public class PostsListSharing extends NewTestTemplate {
         new String[]{"Facebook", "Twitter", "Reddit", "Tumblr"});
   }
 
-  @Test
+  @Test(groups = {"Sharing_002"})
   @Execute(asUser = User.ANONYMOUS)
   @InBrowser(browserSize = DESKTOP_RESOLUTION)
   public void socialNetworkIconsAreDisplayedToAnonymousUserOnDesktop() {
-  toggleShareIconClickDisplaysSocialNetworkIcons(
-      new String[]{"Facebook", "Twitter", "Reddit", "Tumblr"});}
+    toggleShareIconClickDisplaysSocialNetworkIcons(
+        new String[]{"Facebook", "Twitter", "Reddit", "Tumblr"});}
+
+  @Test(groups = {"Sharing_001"})
+  @Execute(asUser = User.ANONYMOUS)
+  @InBrowser(browserSize = DESKTOP_RESOLUTION)
+  public void socialNetworkIconsAreDisplayedToAnonymousUser() {
+    socialNetworkIconsAreDisplayed(new String[]{"Facebook", "Twitter", "Reddit", "Tumblr"});}
 
   /**
-   * LOGGED IN USER SECTION
+   * LOGGED IN USER
    */
-  @Test
+
+  @Test(groups = {"Sharing_002"})
+  @Execute(asUser = User.USER_3)
+  @InBrowser(browserSize = DESKTOP_RESOLUTION)
+  public void socialNetworkIconsAreDisplayedToLoggedInUser() {
+    socialNetworkIconsAreDisplayed(new String[]{"Facebook", "Twitter", "Reddit", "Tumblr"});
+  }
+
+@Test(groups = {"Sharing_003"})
   @Execute(asUser = User.USER_3)
   @InBrowser(browserSize = MOBILE_RESOLUTION)
   public void socialNetworkIconsAreDisplayedToLoggedInUserOnMobile() {
-  toggleShareIconClickDisplaysSocialNetworkIcons(
-      new String[]{"Facebook", "Twitter", "Reddit", "Tumblr"});
+    toggleShareIconClickDisplaysSocialNetworkIcons(
+        new String[]{"Facebook", "Twitter", "Reddit", "Tumblr"});
   }
 
-  @Test
+  @Test(groups = {"Sharing_004"})
   @Execute(asUser = User.USER_3)
   @InBrowser(browserSize = DESKTOP_RESOLUTION)
   public void socialNetworkIconsAreDisplayedToLoggedInUserOnDesktop() {
@@ -81,5 +91,17 @@ public class PostsListSharing extends NewTestTemplate {
     }
   }
 
+  public void socialNetworkIconsAreDisplayed(String[] expectedSocialNetworks) {
+    PostDetailsPage postDetails = new PostDetailsPage(driver).open();
+    String[] currentSocialNetworks = postDetails.getSocialNetworkIconsClasses();
+    for (int i = 0; i < expectedSocialNetworks.length; i++) {
+      String currentSocialNetwork = currentSocialNetworks[i];
+      String expectedSocialNetwork = expectedSocialNetworks[i];
+      Assertion.assertEqualsIgnoreCase(
+          currentSocialNetwork, expectedSocialNetwork,
+          "Expected network not found on its position. Note that the order of social buttons "
+          + "matters, as defined in requirements. Missing network:" + expectedSocialNetwork
+      );
+    }
+  }
 }
-
