+++ src/test/java/com/wikia/webdriver/testcases/discussions/Upvoting.java
@@ -6,41 +6,43 @@ import com.wikia.webdriver.common.core.annotations.InBrowser;
 import com.wikia.webdriver.common.core.helpers.User;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.discussions.PostDetailsPage;
+import com.wikia.webdriver.pageobjectsfactory.pageobject.discussions.PostsListPage;
 
 import org.testng.annotations.Test;
 
-@Test(groups = {"Discussions", "PostDetailsUpvote"})
-public class PostDetailsUpvote extends NewTestTemplate {
+@Test(groups = {"Discussions, Upvoting"})
+public class Upvoting extends NewTestTemplate {
 
   private static final String DESKTOP_RESOLUTION = "1366x768";
   private static final String MOBILE_RESOLUTION = "600x800";
 
   /**
-   * ANONYMOUS USER SECTION
+   * ANONYMOUS USER
+   * POST DETAILS
    */
 
-  @Test
+  @Test(groups = {"Upvoting_001"})
   @Execute(asUser = User.ANONYMOUS)
   @InBrowser(browserSize = MOBILE_RESOLUTION)
   public void AnonymousUserOnMobileCanNotVoteForPostDetails() {
     postDetailsUpvoteButtonClickDoenstAddAnUpvote();
   }
 
-  @Test
+  @Test(groups = {"Upvoting_002"})
   @Execute(asUser = User.ANONYMOUS)
   @InBrowser(browserSize = DESKTOP_RESOLUTION)
   public void AnonymousUserOnDesktopCanNotVoteForPostDetails() {
     postDetailsUpvoteButtonClickDoenstAddAnUpvote();
   }
 
-  @Test
+  @Test(groups = {"Upvoting_003"})
   @Execute(asUser = User.ANONYMOUS)
   @InBrowser(browserSize = MOBILE_RESOLUTION)
   public void AnonymousUserOnMobileCanNotVoteForFirstReply() {
     firstReplyUpvoteButtonClickDoenstAddAnUpvote();
   }
 
-  @Test
+  @Test(groups = {"Upvoting_004"})
   @Execute(asUser = User.ANONYMOUS)
   @InBrowser(browserSize = DESKTOP_RESOLUTION)
   public void AnonymousUserOnDesktopCanNotVoteForFirstReply() {
@@ -48,37 +50,51 @@ public class PostDetailsUpvote extends NewTestTemplate {
   }
 
   /**
+   * ANONYMOUS USER
+   * POST LIST
+   */
+
+  @Test(groups = {"Upvoting_005"})
+  @Execute(asUser = User.ANONYMOUS)
+  @InBrowser(browserSize = MOBILE_RESOLUTION)
+  public void AnonymousUserOnMobileCanNotUpvote() {
+    postListUpvoteButtonClickDoenstAddAnUpvote();
+  }
+
+  @Test(groups = {"Upvoting_006"})
+  @Execute(asUser = User.ANONYMOUS)
+  @InBrowser(browserSize = DESKTOP_RESOLUTION)
+  public void AnonymousUserOnDesktopCanNotUpvote() {
+    postListUpvoteButtonClickDoenstAddAnUpvote();
+  }
+
+  /**
    * LOGGED IN USER
    * POST DETAILS
    */
 
-  @Test
+  @Test(groups = {"Upvoting_007"})
   @Execute(asUser = User.USER_3)
   @InBrowser(browserSize = MOBILE_RESOLUTION)
   public void loggedInUserOnMobileCanSeePostDetailsList() {
     postDetailsUpvoteButtonClickAddsAnUpvoteAndSecondClickRemovesTheUpvote();
   }
 
-  @Test
+  @Test(groups = {"Upvoting_008"})
   @Execute(asUser = User.USER_3)
   @InBrowser(browserSize = DESKTOP_RESOLUTION)
   public void loggedInUserOnDesktopCanSeePostDetailsList() {
     postDetailsUpvoteButtonClickAddsAnUpvoteAndSecondClickRemovesTheUpvote();
   }
 
-  /**
-   * LOGGED IN USER
-   * FIRST REPLY
-   */
-
-  @Test
+  @Test(groups = {"Upvoting_009"})
   @Execute(asUser = User.USER_3)
   @InBrowser(browserSize = MOBILE_RESOLUTION)
   public void loggedInUserOnMobileCanVoteForFirstReply() {
     firstReplyUpvoteButtonClickAddsAnUpvoteAndSecondClickRemovesTheUpvote();
   }
 
-  @Test
+  @Test(groups = {"Upvoting_010"})
   @Execute(asUser = User.USER_3)
   @InBrowser(browserSize = DESKTOP_RESOLUTION)
   public void loggedInUserOnDesktopCanVoteForFirstReply() {
@@ -86,8 +102,53 @@ public class PostDetailsUpvote extends NewTestTemplate {
   }
 
   /**
+   * LOGGED IN USER
+   * POST LIST
+   */
+
+  @Test(groups = {"Upvoting_011"})
+  @Execute(asUser = User.USER_3)
+  @InBrowser(browserSize = MOBILE_RESOLUTION)
+  public void loggedInUserOnMobileCanUpvote() {
+    postListUpvoteButtonClickAddsAnUpvoteAndSecondClickRemovesTheUpvote();
+  }
+
+  @Test(groups = {"Upvoting_012"})
+  @Execute(asUser = User.USER_3)
+  @InBrowser(browserSize = DESKTOP_RESOLUTION)
+  public void loggedInUserOnDesktopCanUpvote() {
+    postListUpvoteButtonClickAddsAnUpvoteAndSecondClickRemovesTheUpvote();
+  }
+
+  /**
    * TESTING METHODS SECTION
    */
+
+  public void postListUpvoteButtonClickAddsAnUpvoteAndSecondClickRemovesTheUpvote() {
+    PostsListPage postList = new PostsListPage(driver).open();
+    int replyIndex = 0;
+    postList.isUpvoteButtonVisible(replyIndex);
+    String firstVoteCount = postList.getVoteCount(replyIndex);
+    postList.clickUpvoteButton(replyIndex);
+    postList.waitForVoteCountToChange(replyIndex, firstVoteCount);
+    String secondVoteCount = postList.getVoteCount(replyIndex);
+    Assertion.assertNotEquals(firstVoteCount, secondVoteCount);
+    postList.clickUpvoteButton(replyIndex);
+    postList.waitForVoteCountToChange(replyIndex, secondVoteCount);
+    String thirdVoteCount = postList.getVoteCount(replyIndex);
+    Assertion.assertEquals(firstVoteCount, thirdVoteCount);
+  }
+
+  public void postListUpvoteButtonClickDoenstAddAnUpvote() {
+    PostsListPage postList = new PostsListPage(driver).open();
+    int replyIndex = 0;
+    postList.isUpvoteButtonVisible(replyIndex);
+    String firstVoteCount = postList.getVoteCount(replyIndex);
+    postList.clickUpvoteButton(replyIndex);
+    postList.waitForVoteCountChangeTimeLagToPass();
+    String secondVoteCount = postList.getVoteCount(replyIndex);
+    Assertion.assertEquals(firstVoteCount, secondVoteCount);
+  }
 
   public void postDetailsUpvoteButtonClickAddsAnUpvoteAndSecondClickRemovesTheUpvote() {
     PostDetailsPage postDetails = new PostDetailsPage(driver).open();
